+++ hfut/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding:utf-8 -*-
+"""
+合肥工业大学教务系统学生端接口以及方便开发者开发围绕学生数据的一些工具.
+"""
+from __future__ import unicode_literals
+
+__title__ = 'hfut'
+__version__ = '1.4.0'
+__author__ = 'erliang'
+__author_email__ = 'dev@erliang.me'
+__url__ = 'https://github.com/er1iang/hfut'
+__license__ = 'MIT'
+__copyright__ = 'Copyright 2015-2016 erliang'
+
+from .log import *
+from .model import *
+from .value import *
+from .exception import *